+++ gradio/__init__.py
@@ -121,6 +121,7 @@ from gradio.oauth import OAuthProfile, OAuthToken
 from gradio.renderable import render
 from gradio.route_utils import Header
 from gradio.routes import Request, mount_gradio_app
+from gradio.server import Server
 from gradio.templates import (
     Files,
     ImageMask,
@@ -143,6 +144,7 @@ __version__ = get_package_version()
 __all__ = [
     "Accordion",
     "AnnotatedImage",
+    "Server",
     "Annotatedimage",
     "Audio",
     "BarPlot",