+++ js/checkbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/checkbox
 
+## 0.6.5
+
+### Dependency updates
+
+- @gradio/utils@0.12.1
+- @gradio/statustracker@0.13.0
+
 ## 0.6.4
 
 ### Dependency updates